--- gen_tso/last_updated_trexolist.txt
@@ -1,1 +1,0 @@
-2025_11_06
